- Added more DP helpers - Added more enumerable helpers - Added more vtable data - Move `Infiltrate` hook - Move more functions to `.cpp` to reduce clutter

+++ YRpp/NetworkEvents.h
@@ -66,7 +66,7 @@ struct NetworkEvent {
 	NetworkEvent * FillEvent_SWPlace(int PlayerNumber, NetworkEvents eventKind, int swTypeIdx, CellStruct *loc)
 		{ JMP_THIS(0x4C6B60); }
 
-	static constexpr reference<int, 0x8208ECu , 36u> const EventLength{};
+	static constexpr reference<size_t, 0x8208ECu , 36u> const EventLength{};
 };
 
 // This Replaced with TargetClass
